+++ boj/2023/tiers/bronze/15596.cpp
@@ -1,8 +1,8 @@
 #include <vector>
 
-long long sum(std::vector<int> &a)
+long sum(std::vector<int> &a)
 {
-	long long ret = 0;
+	long ret = 0;
 	for (int x : a)
 		ret += x;
 	return ret;
